Made slight modifications to the Specifications table on lines 9 ~ 15.
Changed the original content under Design Considerations and to Under Development.

Under Module Pin Configuration, the following changes were made:
Line 48 ~ Bootstrap A was modified to Module Specific with the corresponding function.
Lines 55 and 56 ~ Ethernet LED SPEED and LINK were changed to Indicator A and Indicator B respectively.
Lines 134 and 135 ~ A backslash was added to the left of the pipe between USART6 TX | T3C1 and USART6 RX | T3C2 so the pipe will show up in the table on the web page.
Line 152 ~ Module Specific 1 was changed to Bootstrap A with the corresponding function.
Line 212 ~ Reserved was modified to Indicator C with the corresponding function.
Line 245 ~ Reserved was modified to Indicator D with the corresponding function.