Expanded the Specification table to include a third column.
Added a header for the second and third columns on line 9 (UC5550P40 and UC5550P80 respectively).

Made a few changes to the second column after the third column was added as followed:
Line 12 ~ Changed 100 MHz to 180 MHz.
Line 13 ~ Changed 320 KB (SRAM) to 256 KB (SRAM).
Line 14 ~ Added a space between the 1 and MB.
Line 15 ~ (pushed to line 16) Changed 4MB (QSPI) to Not supported.

Added a new row to the table @line 15 for External RAM

Under Module Pin Configuration, the Function Name on line 55 was changed to dashes.