PWM Fix
There is a bug in PWM dealing with a possible timer overflow where you need to call Stop(), change frequency, and then call Start() again to allow the PWM to work. This was discovered while using the Musical Tones section. I am not sure if the problem will present itself with the other examples.